# api/openapi.yaml
openapi: 3.0.3
info:
  title: Village Calendar Service REST API
  version: 1.0.0
  description: REST API for webhooks, OAuth2 authentication, and health monitoring. For calendar operations, use the GraphQL API at /graphql.
  contact:
    name: VillageCompute Support
    url: https://villagecompute.com
servers:
  - url: http://localhost:8030
    description: Development server
  - url: https://calendar.villagecompute.com
    description: Production server
...
paths:
...
  /auth/login/google:
    get:
      tags:
        - Authentication
      summary: Initiate Google OAuth login
      description: |
        Redirects user to Google OAuth consent page. This endpoint triggers the OAuth2
        authorization code flow. After user grants permissions on Google's consent page,
        they will be redirected back to /auth/google/callback with an authorization code.
      operationId: loginWithGoogle
      responses:
        '303':
          description: Redirect to Google OAuth consent page or callback endpoint
        '401':
          description: Authentication required or OAuth flow failed
  /auth/google/callback:
    get:
      tags:
        - Authentication
      summary: Handle Google OAuth callback
      description: |
        OAuth2 callback endpoint invoked by Google after user grants permissions.
        Exchanges the authorization code for an access token, creates/updates the user record,
        and issues a JWT token for subsequent API requests. This endpoint is NOT directly callable
        by clients - it's part of the OAuth2 authorization code flow.
      operationId: handleGoogleCallback
      responses:
        '200':
          description: Authentication successful, JWT token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "user@example.com"
                  displayName: "John Doe"
                  profileImageUrl: "https://example.com/avatar.jpg"
        '500':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Authentication failed: ..."
  /auth/login/facebook:
    get:
      tags:
        - Authentication
      summary: Initiate Facebook OAuth login
      description: |
        Redirects user to Facebook OAuth consent page. This endpoint triggers the OAuth2
        authorization code flow. After user grants permissions on Facebook's consent page,
        they will be redirected back to /auth/facebook/callback with an authorization code.
      operationId: loginWithFacebook
      responses:
        '303':
          description: Redirect to Facebook OAuth consent page or callback endpoint
        '401':
          description: Authentication required or OAuth flow failed
  /auth/facebook/callback:
    get:
      tags:
        - Authentication
      summary: Handle Facebook OAuth callback
      description: |
        OAuth2 callback endpoint invoked by Facebook after user grants permissions.
        Exchanges the authorization code for an access token, creates/updates the user record,
        and issues a JWT token for subsequent API requests. This endpoint is NOT directly callable
        by clients - it's part of the OAuth2 authorization code flow.
      operationId: handleFacebookCallback
      responses:
        '200':
          description: Authentication successful, JWT token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "user@example.com"
                  displayName: "Jane Smith"
                  profileImageUrl: "https://example.com/avatar.jpg"
        '500':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Authentication failed: ..."
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current authenticated user
      description: |
        Returns user information based on the JWT token provided in the Authorization header.
        Used to verify token validity and retrieve current user details.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                email: "user@example.com"
                displayName: "John Doe"
                profileImageUrl: "https://example.com/avatar.jpg"
        '401':
          description: Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "No valid JWT token found"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "User not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Failed to fetch user: ..."
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from OAuth2 callback endpoints
  schemas:
...
    AuthResponse:
      type: object
      description: Authentication response containing JWT token and user information
      properties:
        token:
          type: string
          description: JWT token for API authentication
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      description: User information
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier (UUID)
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        displayName:
          type: string
          description: User display name
          example: "John Doe"
        profileImageUrl:
          type: string
          format: uri
          description: URL to user's profile image
          example: "https://example.com/avatar.jpg"
    ErrorResponse:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error message
          example: "Authentication failed"
...